server.js split into modules
